+++ BranchSDK/src/BranchIO/Internal/IClientSession.h
